GERONIMO-2087 MimeUtility.encodeWord()/encodeText() have some errors.
git-svn-id: https://svn.apache.org/repos/asf/geronimo/specs/trunk/geronimo-spec-javamail@412426 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/geronimo/mail/util/Base64Encoder.java
@@ -515,21 +515,22 @@ public class Base64Encoder
             int readCount = in.read(inBuffer);
             // did we get a full triplet?  that's an easy encoding.
             if (readCount == 3) {
-                byte a1 = (byte)(inBuffer[0] & 0xff);
-                byte a2 = (byte)(inBuffer[1] & 0xff);
-                byte a3 = (byte)(inBuffer[2] & 0xff);
+                int  a1 = inBuffer[0] & 0xff;
+                int  a2 = inBuffer[1] & 0xff;
+                int  a3 = inBuffer[2] & 0xff;
 
                 out.append((char)encodingTable[(a1 >>> 2) & 0x3f]);
                 out.append((char)encodingTable[((a1 << 4) | (a2 >>> 4)) & 0x3f]);
                 out.append((char)encodingTable[((a2 << 2) | (a3 >>> 6)) & 0x3f]);
                 out.append((char)encodingTable[a3 & 0x3f]);
+
             }
             else if (readCount <= 0) {
                 // eof condition, don'e entirely.
                 return;
             }
             else if (readCount == 1) {
-                byte a1 = (byte)(inBuffer[0] & 0xff);
+                int  a1 = inBuffer[0] & 0xff;
                 out.append((char)encodingTable[(a1 >>> 2) & 0x3f]);
                 out.append((char)encodingTable[(a1 << 4) & 0x3f]);
                 out.append((char)padding);
@@ -537,8 +538,8 @@ public class Base64Encoder
                 return;
             }
             else if (readCount == 2) {
-                byte a1 = (byte)(inBuffer[0] & 0xff);
-                byte a2 = (byte)(inBuffer[1] & 0xff);
+                int  a1 = inBuffer[0] & 0xff;
+                int  a2 = inBuffer[1] & 0xff;
 
                 out.append((char)encodingTable[(a1 >>> 2) & 0x3f]);
                 out.append((char)encodingTable[((a1 << 4) | (a2 >>> 4)) & 0x3f]);
